requirements.txt
Flask>=2.0
python-dotenv>=0.15
openai>=1.0
PyPDF2>=3.0
python-docx>=1.0
Werkzeug>=2.0
chromadb>=0.4.15
